requirements.txt
langgraph==0.2.34
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-openai==0.2.1
openai
python-dotenv==1.0.1
pydantic==2.9.2
typing-extensions==4.12.2
asyncio
